+++ kernel/extended/liteipc/hm_liteipc.c
@@ -1216,12 +1216,12 @@ LITE_OS_SEC_TEXT int LiteIpcIoctl(struct file *filep, int cmd, unsigned long arg
             }
             ret = LiteIpcMsgHandle((IpcContent *)(UINTPTR)arg);
             if (ret != LOS_OK) {
-                return ret;
+                return (int)ret;
             }
             break;
         default:
             PRINT_ERR("Unknow liteipc ioctl cmd:%d\n", cmd);
             return -EINVAL;
     }
-    return ret;
+    return (int)ret;
 }
